+++ Assets/MyProject/Scripts/Characters/BulletController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cea912bca8b10ba46b39b5905857b118
+guid: 9ddedd1adc0eb9a4399b342321adf6f2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
